Switch to using Nu Validator directly
Previously, we were invoking the Nu Validator through
[html5validator](https://github.com/svenkreiss/html5validator), but that
project hasn't been updated since May 2022, so it's using an old version
of the Nu Validator, which results in false positive error messages.

This patch switches to using the latest version of Nu Validator through
the vnu.jar command-line interface. Because the [Nu Validator
project](https://github.com/validator/validator) hasn't tagged and
released a recent version, and because we don't want to rely on the
untagged "latest" release, we build the validator from source, using a
pinned Git commit hash.